renames file to be consistent with most of the repo
