+++ packages/relay/src/lib/services/index.ts
@@ -7,7 +7,7 @@ import { ICommonService } from './ethService/ethCommonService/ICommonService';
 import { FilterService } from './ethService/ethFilterService';
 import { IFilterService } from './ethService/ethFilterService/IFilterService';
 import MetricService from './metricService/metricService';
-import { TransactionService } from './transactionService/TransactionService';
+import { TransactionService } from './transactionService/transactionService';
 
 export {
   BlockService,
